Switch to manual prometheus operator deployment.
Instead of relying on OLM to provision the prometheus operator, we are
opting to manually deploy the prometheus operator. This reduces a lot of
overhead involved with adding a new olm dependency within rhods. It will
also give us a lot more flexibility in the sort of namespaces the
prometheus operator has access to, without conflicting with existing
ocp monitoring stack.

Signed-off-by: Humair Khan <humair88@hotmail.com>