Leverage upstream servicemonitors for modelmesh.
Now that the prometheus operator is deployed without olm, we can have
the operator watch for servicemonitors in other namespaces without
conflicting with the ocp monitoring prometheus operator. As such, we do
not need to deploy our own service monitor, and instead we can have this
prometheus operator observe service monitors deployed by modelmesh
controller.

The only requirement is that the servicemonitors have the designated
modelmesh label, along with a rolebinding that gives the prometheus SA
the "rhods-prometheus-k8s" cluster role permissions.

Signed-off-by: Humair Khan <humair88@hotmail.com>